Poprawki

+++ php_scripts/user_panel/delete_offer.php
@@ -11,18 +11,20 @@ try{//usuwanie zdjęć trzeba tu zrobić
     require_once '../connect.php';
     $connection = mysqli_connect($host,$db_user,$db_password,$db_name);
     if($connection->connect_errno==0){
-        $sql="DELETE FROM users_offers WHERE seller='$user' AND offer_id='$offer' AND status='available'";
-        $result=mysqli_query($connection,$sql);
-        $result=mysqli_affected_rows($connection);
-        if($result){
-            $sql="SELECT photo1, photo2 FROM users_offers WHERE seller='$user' AND offer_id='$offer'";
-            $result = mysqli_query($connection,$sql); 
-            $result=mysqli_fetch_row($result);
-            foreach($result as $photo){
-                if($photo){
-                    $path="../../$photo";
-                    if(file_exists($path)){
-                        unlink($path);
+        $sql_img="SELECT photo1, photo2 FROM users_offers WHERE seller='$user' AND offer_id='$offer'";
+        $result_img = mysqli_query($connection,$sql_img); 
+        $result_img=mysqli_fetch_row($result_img);
+        if($result_img){
+            $sql="DELETE FROM users_offers WHERE seller='$user' AND offer_id='$offer' AND status='available'";
+            $result=mysqli_query($connection,$sql);
+            $result=mysqli_affected_rows($connection);
+            if($result){
+                foreach($result_img as $photo){
+                    if($photo){
+                        $path="../../$photo";
+                        if(file_exists($path)){
+                            unlink($path);
+                        }
                     }
                 }
             }
